Add more custom fake message, random fake message
